more dtype fixes

+++ tests/augmentation/test_random_generator.py
@@ -195,7 +195,7 @@ class TestColorJiggleGen(RandomGeneratorBaseTests):
                 device=device,
                 dtype=dtype,
             ),
-            "order": torch.tensor([3, 2, 0, 1], device=device, dtype=dtype),
+            "order": torch.tensor([3, 2, 0, 1], device=device, dtype=torch.long),
         }
 
         assert set(jitter_params.keys()) == {
@@ -232,7 +232,7 @@ class TestColorJiggleGen(RandomGeneratorBaseTests):
             atol=1e-4,
         )
         assert_close(
-            jitter_params["order"].to(dtype),
+            jitter_params["order"],
             expected_jitter_params["order"],
             rtol=1e-4,
             atol=1e-4,
@@ -269,7 +269,7 @@ class TestColorJiggleGen(RandomGeneratorBaseTests):
                 device=device,
                 dtype=dtype,
             ),
-            "order": torch.tensor([3, 2, 0, 1], device=device, dtype=dtype),
+            "order": torch.tensor([3, 2, 0, 1], device=device, dtype=torch.long),
         }
 
         assert set(jitter_params.keys()) == {
@@ -306,7 +306,7 @@ class TestColorJiggleGen(RandomGeneratorBaseTests):
             atol=1e-4,
         )
         assert_close(
-            jitter_params["order"].to(dtype),
+            jitter_params["order"],
             expected_jitter_params["order"],
             rtol=1e-4,
             atol=1e-4,
@@ -327,7 +327,7 @@ class TestColorJiggleGen(RandomGeneratorBaseTests):
             "contrast_factor": torch.tensor([1.2490] * batch_size, device=device, dtype=dtype),
             "hue_factor": torch.tensor([-0.0234] * batch_size, device=device, dtype=dtype),
             "saturation_factor": torch.tensor([1.3674] * batch_size, device=device, dtype=dtype),
-            "order": torch.tensor([2, 3, 0, 1], device=device, dtype=dtype),
+            "order": torch.tensor([2, 3, 0, 1], device=device, dtype=torch.long),
         }
 
         assert_close(
